+++ packages/core/test/prompts/password.test.ts
@@ -29,6 +29,18 @@ describe('PasswordPrompt', () => {
 		expect(output.buffer).to.deep.equal([cursor.hide, 'foo']);
 	});
 
+	test('returns empty string on empty submit', async () => {
+		const instance = new PasswordPrompt({
+			input,
+			output,
+			render: () => 'foo',
+		});
+		const resultPromise = instance.prompt();
+		input.emit('keypress', '', { name: 'return' });
+		const result = await resultPromise;
+		expect(result).to.equal('');
+	});
+
 	describe('cursor', () => {
 		test('can get cursor', () => {
 			const instance = new PasswordPrompt({